Initial work on pipes(Functional version)
