+++ input.txt
@@ -1,0 +1,18 @@
+LICENSE
+Makefile
+README.MD
+cmd
+exit_status_tester
+exit_status_tester.c
+include
+input.txt
+libs
+main.c
+main.o
+minishell
+output.txt
+parser
+pipe
+redirection
+signal.c
+signal.o
