+++ redirection/utils_redir.c
@@ -31,23 +31,3 @@ int	redir_chk(char **args)
 	return (0);
 }
 
-void	exec_child(char **args, char **envp, int fds[], int *exit_status)
-{
-	(void)envp;
-	(void)args;
-	dup2(fds[0], 0);
-	close(fds[0]);
-	close(fds[1]);
-	// execute
-	exit(*exit_status);
-}
-
-void	exec_parents(int pid, int fds[], int *exit_status)
-{
-	int	status;
-
-	close(fds[1]);
-	close(fds[0]);
-	waitpid(pid, &status, 0);
-	*exit_status = status >> 8;
-}
